+++ flink-yarn/src/main/java/org/apache/flink/yarn/YarnFlinkApplicationMasterRunner.java
@@ -23,24 +23,17 @@ import org.apache.flink.api.common.JobExecutionResult;
 import org.apache.flink.api.common.time.Time;
 import org.apache.flink.configuration.ConfigConstants;
 import org.apache.flink.configuration.Configuration;
-import org.apache.flink.configuration.GlobalConfiguration;
-import org.apache.flink.configuration.HighAvailabilityOptions;
 import org.apache.flink.runtime.akka.AkkaUtils;
 import org.apache.flink.runtime.client.JobExecutionException;
 import org.apache.flink.runtime.clusterframework.ApplicationStatus;
 import org.apache.flink.runtime.clusterframework.BootstrapTools;
-import org.apache.flink.runtime.execution.librarycache.BlobLibraryCacheManager;
 import org.apache.flink.runtime.highavailability.HighAvailabilityServices;
 import org.apache.flink.runtime.highavailability.HighAvailabilityServicesUtils;
 import org.apache.flink.runtime.jobgraph.JobGraph;
 import org.apache.flink.runtime.jobmanager.OnCompletionActions;
-import org.apache.flink.runtime.jobmaster.JobManagerServices;
-import org.apache.flink.runtime.jobmaster.JobMaster;
-import org.apache.flink.runtime.leaderelection.LeaderContender;
-import org.apache.flink.runtime.leaderelection.LeaderElectionService;
+import org.apache.flink.runtime.jobmaster.JobManagerRunner;
 import org.apache.flink.runtime.metrics.MetricRegistry;
 import org.apache.flink.runtime.metrics.MetricRegistryConfiguration;
-import org.apache.flink.runtime.metrics.groups.JobManagerMetricGroup;
 import org.apache.flink.runtime.resourcemanager.JobLeaderIdService;
 import org.apache.flink.runtime.resourcemanager.ResourceManager;
 import org.apache.flink.runtime.resourcemanager.ResourceManagerConfiguration;
@@ -50,14 +43,9 @@ import org.apache.flink.runtime.resourcemanager.slotmanager.SlotManagerFactory;
 import org.apache.flink.runtime.rpc.FatalErrorHandler;
 import org.apache.flink.runtime.rpc.RpcService;
 import org.apache.flink.runtime.rpc.akka.AkkaRpcService;
-import org.apache.flink.runtime.security.SecurityContext;
 import org.apache.flink.runtime.util.EnvironmentInformation;
 import org.apache.flink.runtime.util.JvmShutdownSafeguard;
 import org.apache.flink.runtime.util.SignalHandler;
-import org.apache.flink.yarn.cli.FlinkYarnSessionCli;
-import org.apache.hadoop.fs.CommonConfigurationKeysPublic;
-import org.apache.hadoop.security.UserGroupInformation;
-import org.apache.hadoop.yarn.api.ApplicationConstants.Environment;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import scala.concurrent.duration.FiniteDuration;
@@ -66,29 +54,22 @@ import javax.annotation.concurrent.GuardedBy;
 
 import java.io.File;
 import java.io.FileInputStream;
-import java.io.IOException;
 import java.io.ObjectInputStream;
-import java.util.Map;
-import java.util.UUID;
 
 /**
  * This class is the executable entry point for the YARN application master.
- * It starts actor system and the actors for {@link org.apache.flink.runtime.jobmaster.JobMaster}
+ * It starts actor system and the actors for {@link org.apache.flink.runtime.jobmaster.JobManagerRunner}
  * and {@link org.apache.flink.yarn.YarnResourceManager}.
  *
- * The JobMasters handles Flink job execution, while the YarnResourceManager handles container
+ * The JobMasnagerRunner start a {@link org.apache.flink.runtime.jobmaster.JobMaster}
+ * JobMaster handles Flink job execution, while the YarnResourceManager handles container
  * allocation and failure detection.
  */
-public class YarnFlinkApplicationMasterRunner implements LeaderContender, OnCompletionActions, FatalErrorHandler {
+public class YarnFlinkApplicationMasterRunner extends AbstractYarnFlinkApplicationMasterRunner
+		implements OnCompletionActions, FatalErrorHandler {
 
 	/** Logger */
 	protected static final Logger LOG = LoggerFactory.getLogger(YarnFlinkApplicationMasterRunner.class);
-
-	/** The process environment variables */
-	private static final Map<String, String> ENV = System.getenv();
-
-	/** The exit code returned if the initialization of the application master failed */
-	private static final int INIT_ERROR_EXIT_CODE = 31;
 
 	/** The job graph file path */
 	private static final String JOB_GRAPH_FILE_PATH = "flink.jobgraph.path";
@@ -103,31 +84,17 @@ public class YarnFlinkApplicationMasterRunner implements LeaderContender, OnComp
 	private HighAvailabilityServices haServices;
 
 	@GuardedBy("lock")
-	private LeaderElectionService jmLeaderElectionService;
-
-	@GuardedBy("lock")
-	private RpcService jobMasterRpcService;
-
-	@GuardedBy("lock")
-	private RpcService resourceManagerRpcService;
+	private RpcService commonRpcService;
 
 	@GuardedBy("lock")
 	private ResourceManager resourceManager;
 
 	@GuardedBy("lock")
-	private JobMaster jobMaster;
-
-	@GuardedBy("lock")
-	JobManagerServices jobManagerServices;
-
-	@GuardedBy("lock")
-	JobManagerMetricGroup jobManagerMetrics;
+	private JobManagerRunner jobManagerRunner;
 
 	@GuardedBy("lock")
 	private JobGraph jobGraph;
 
-	/** Flag marking the app master runner as started/running */
-	private volatile boolean running;
 	// ------------------------------------------------------------------------
 	//  Program entry point
 	// ------------------------------------------------------------------------
@@ -147,105 +114,11 @@ public class YarnFlinkApplicationMasterRunner implements LeaderContender, OnComp
 		System.exit(returnCode);
 	}
 
-	/**
-	 * The instance entry point for the YARN application master. Obtains user group
-	 * information and calls the main work method {@link #runApplicationMaster(org.apache.flink.configuration.Configuration)} as a
-	 * privileged action.
-	 *
-	 * @param args The command line arguments.
-	 * @return The process exit code.
-	 */
-	protected int run(String[] args) {
-		try {
-			LOG.debug("All environment variables: {}", ENV);
-
-			final String yarnClientUsername = ENV.get(YarnConfigKeys.ENV_HADOOP_USER_NAME);
-			require(yarnClientUsername != null, "YARN client user name environment variable {} not set",
-				YarnConfigKeys.ENV_HADOOP_USER_NAME);
-
-			final String currDir = ENV.get(Environment.PWD.key());
-			require(currDir != null, "Current working directory variable (%s) not set", Environment.PWD.key());
-			LOG.debug("Current working Directory: {}", currDir);
-
-			final String remoteKeytabPath = ENV.get(YarnConfigKeys.KEYTAB_PATH);
-			LOG.debug("remoteKeytabPath obtained {}", remoteKeytabPath);
-
-			final String remoteKeytabPrincipal = ENV.get(YarnConfigKeys.KEYTAB_PRINCIPAL);
-			LOG.info("remoteKeytabPrincipal obtained {}", remoteKeytabPrincipal);
-
-			String keytabPath = null;
-			if(remoteKeytabPath != null) {
-				File f = new File(currDir, Utils.KEYTAB_FILE_NAME);
-				keytabPath = f.getAbsolutePath();
-				LOG.debug("keytabPath: {}", keytabPath);
-			}
-
-			UserGroupInformation currentUser = UserGroupInformation.getCurrentUser();
-
-			LOG.info("YARN daemon is running as: {} Yarn client user obtainer: {}",
-					currentUser.getShortUserName(), yarnClientUsername );
-
-			SecurityContext.SecurityConfiguration sc = new SecurityContext.SecurityConfiguration();
-
-			//To support Yarn Secure Integration Test Scenario
-			File krb5Conf = new File(currDir, Utils.KRB5_FILE_NAME);
-			if(krb5Conf.exists() && krb5Conf.canRead()) {
-				String krb5Path = krb5Conf.getAbsolutePath();
-				LOG.info("KRB5 Conf: {}", krb5Path);
-				org.apache.hadoop.conf.Configuration conf = new org.apache.hadoop.conf.Configuration();
-				conf.set(CommonConfigurationKeysPublic.HADOOP_SECURITY_AUTHENTICATION, "kerberos");
-				conf.set(CommonConfigurationKeysPublic.HADOOP_SECURITY_AUTHORIZATION, "true");
-				sc.setHadoopConfiguration(conf);
-			}
-
-			// Flink configuration
-			final Map<String, String> dynamicProperties =
-					FlinkYarnSessionCli.getDynamicProperties(ENV.get(YarnConfigKeys.ENV_DYNAMIC_PROPERTIES));
-			LOG.debug("YARN dynamic properties: {}", dynamicProperties);
-
-			final Configuration flinkConfig = createConfiguration(currDir, dynamicProperties);
-			if(keytabPath != null && remoteKeytabPrincipal != null) {
-				flinkConfig.setString(ConfigConstants.SECURITY_KEYTAB_KEY, keytabPath);
-				flinkConfig.setString(ConfigConstants.SECURITY_PRINCIPAL_KEY, remoteKeytabPrincipal);
-			}
-
-			SecurityContext.install(sc.setFlinkConfiguration(flinkConfig));
-
-			return SecurityContext.getInstalled().runSecured(new SecurityContext.FlinkSecuredRunner<Integer>() {
-				@Override
-				public Integer run() {
-					return runApplicationMaster(flinkConfig);
-				}
-			});
-
-		}
-		catch (Throwable t) {
-			// make sure that everything whatever ends up in the log
-			LOG.error("YARN Application Master initialization failed", t);
-			return INIT_ERROR_EXIT_CODE;
-		}
-	}
-
-	// ------------------------------------------------------------------------
-	//  Core work method
-	// ------------------------------------------------------------------------
-
-	/**
-	 * The main work method, must run as a privileged action.
-	 *
-	 * @return The return code for the Java process.
-	 */
+	@Override
 	protected int runApplicationMaster(Configuration config) {
 
 		try {
 			// ---- (1) create common services
-			// Note that we use the "appMasterHostname" given by YARN here, to make sure
-			// we use the hostnames given by YARN consistently throughout akka.
-			// for akka "localhost" and "localhost.localdomain" are different actors.
-			final String appMasterHostname = ENV.get(Environment.NM_HOST.key());
-			require(appMasterHostname != null,
-					"ApplicationMaster hostname variable %s not set", Environment.NM_HOST.key());
-			LOG.info("YARN assigned hostname for application master: {}", appMasterHostname);
 
 			// try to start the rpc service
 			// using the port range definition from the config.
@@ -256,38 +129,27 @@ public class YarnFlinkApplicationMasterRunner implements LeaderContender, OnComp
 			synchronized (lock) {
 				haServices = HighAvailabilityServicesUtils.createAvailableOrEmbeddedServices(config);
 				metricRegistry = new MetricRegistry(MetricRegistryConfiguration.fromConfiguration(config));
+				commonRpcService = createRpcService(config, appMasterHostname, amPortRange);
 
 				// ---- (2) init resource manager -------
-				resourceManagerRpcService = createRpcService(config, appMasterHostname, amPortRange);
 				resourceManager = createResourceManager(config);
 
 				// ---- (3) init job master parameters
-				jobMasterRpcService = createRpcService(config, appMasterHostname, amPortRange);
-				jobManagerServices = JobManagerServices.fromConfiguration(config, haServices);
-				jobManagerMetrics = new JobManagerMetricGroup(metricRegistry, jobMasterRpcService.getAddress());
-				jobMaster = createJobMaster(config);
+				jobManagerRunner = createJobManagerRunner(config);
 
-				// ---- (4) start the resource manager  and job master:
+				// ---- (4) start the resource manager  and job manager runner:
 				resourceManager.start();
 				LOG.debug("YARN Flink Resource Manager started");
 
-				// mark the job as running in the HA services
-				try {
-					haServices.getRunningJobsRegistry().setJobRunning(jobGraph.getJobID());
-				}
-				catch (Throwable t) {
-					throw new JobExecutionException(jobGraph.getJobID(),
-							"Could not register the job at the high-availability services", t);
-				}
-				jmLeaderElectionService.start(this);
+				jobManagerRunner.start();
+				LOG.debug("Job Manager Runner started");
 
 				// ---- (5) start the web monitor
 				// TODO: add web monitor
 			}
-			running = true;
-			while (running) {
-				Thread.sleep(100);
-			}
+
+			// wait for resource manager to finish
+			resourceManager.getTerminationFuture().get();
 			// everything started, we can wait until all is done or the process is killed
 			LOG.info("YARN Application Master finished");
 		}
@@ -305,19 +167,6 @@ public class YarnFlinkApplicationMasterRunner implements LeaderContender, OnComp
 	//  Utilities
 	// ------------------------------------------------------------------------
 
-	/**
-	 * Validates a condition, throwing a RuntimeException if the condition is violated.
-	 * 
-	 * @param condition The condition.
-	 * @param message The message for the runtime exception, with format variables as defined by
-	 *				{@link String#format(String, Object...)}.
-	 * @param values The format arguments.
-	 */
-	private static void require(boolean condition, String message, Object... values) {
-		if (!condition) {
-			throw new RuntimeException(String.format(message, values));
-		}
-	}
 	protected RpcService createRpcService(
 			Configuration configuration,
 			String bindAddress,
@@ -334,7 +183,7 @@ public class YarnFlinkApplicationMasterRunner implements LeaderContender, OnComp
 
 		return new YarnResourceManager(config,
 				ENV,
-				resourceManagerRpcService,
+				commonRpcService,
 				resourceManagerConfiguration,
 				haServices,
 				slotManagerFactory,
@@ -343,73 +192,36 @@ public class YarnFlinkApplicationMasterRunner implements LeaderContender, OnComp
 				this);
 	}
 
-	private JobMaster createJobMaster(Configuration config) throws Exception{
-		// get JobGraph from local resources
+	private JobManagerRunner createJobManagerRunner(Configuration config) throws Exception{
+		// first get JobGraph from local resources
 		jobGraph = loadJobGraph(config);
-		if (jobGraph == null) {
-			throw new Exception("Fail to load job graph");
-		}
 
-		// libraries and class loader
-		final BlobLibraryCacheManager libraryCacheManager = jobManagerServices.libraryCacheManager;
+		// we first need to mark the job as running in the HA services, so that the
+		// JobManager leader will recognize that it as work to do
 		try {
-			libraryCacheManager.registerJob(
-					jobGraph.getJobID(), jobGraph.getUserJarBlobKeys(), jobGraph.getClasspaths());
-		} catch (IOException e) {
-			throw new Exception("Cannot set up the user code libraries: " + e.getMessage(), e);
+			haServices.getRunningJobsRegistry().setJobRunning(jobGraph.getJobID());
+		}
+		catch (Throwable t) {
+			throw new JobExecutionException(jobGraph.getJobID(),
+					"Could not register the job at the high-availability services", t);
 		}
 
-		final ClassLoader userCodeLoader = libraryCacheManager.getClassLoader(jobGraph.getJobID());
-		if (userCodeLoader == null) {
-			throw new Exception("The user code class loader could not be initialized.");
-		}
-		// set self address to ha service for rm to find itself.
-		jmLeaderElectionService = haServices.getJobManagerLeaderElectionService(jobGraph.getJobID());
-
-		// now the JobManager
-		return new JobMaster(
+		// now the JobManagerRunner
+		return new JobManagerRunner(
 				jobGraph, config,
-				jobMasterRpcService,
+				commonRpcService,
 				haServices,
-				jobManagerServices.executorService,
-				jobManagerServices.libraryCacheManager,
-				jobManagerServices.restartStrategyFactory,
-				jobManagerServices.rpcAskTimeout,
-				jobManagerMetrics,
 				this,
-				this,
-				userCodeLoader);
+				this);
 	}
 
 	protected void shutdown(ApplicationStatus status, String msg) {
 		synchronized (lock) {
-			try {
-				haServices.getRunningJobsRegistry().setJobFinished(jobGraph.getJobID());
-			}
-			catch (Throwable t) {
-				LOG.error("Could not un-register from high-availability services job {} ({}).",
-						jobGraph.getName(), jobGraph.getJobID(), t);
-			}
-			try {
-				jobManagerServices.shutdown();
-			} catch (Throwable tt) {
-				LOG.error("Error while shutting down JobManager services", tt);
-			}
-			if (jobManagerMetrics != null) {
-				jobManagerMetrics.close();
-			}
-			if (jmLeaderElectionService != null) {
+			if (jobManagerRunner != null) {
 				try {
-					jmLeaderElectionService.stop();
-				} catch (Throwable ignored) {
-					LOG.warn("Failed to stop the job master leader election service", ignored);
-				}
-			}
-			if (jobMaster != null) {
-				try {
-					jobMaster.shutDown();
+					jobManagerRunner.shutdown();
 				} catch (Throwable tt) {
-					LOG.warn("Failed to stop the JobMaster", tt);
+					LOG.warn("Failed to stop the JobManagerRunner", tt);
 				}
 			}
 			if (resourceManager != null) {
@@ -420,18 +232,11 @@ public class YarnFlinkApplicationMasterRunner implements LeaderContender, OnComp
 					LOG.warn("Failed to stop the ResourceManager", tt);
 				}
 			}
-			if (resourceManagerRpcService != null) {
+			if (commonRpcService != null) {
 				try {
-					resourceManagerRpcService.stopService();
+					commonRpcService.stopService();
 				} catch (Throwable tt) {
 					LOG.error("Error shutting down resource manager rpc service", tt);
-				}
-			}
-			if (jobMasterRpcService != null) {
-				try {
-					jobMasterRpcService.stopService();
-				} catch (Throwable tt) {
-					LOG.error("Error shutting down job master rpc service", tt);
 				}
 			}
 			if (haServices != null) {
@@ -449,13 +254,11 @@ public class YarnFlinkApplicationMasterRunner implements LeaderContender, OnComp
 				}
 			}
 		}
-		running = false;
 	}
 
 	private static JobGraph loadJobGraph(Configuration config) throws Exception {
-		// TODO:
 		JobGraph jg = null;
-		String jobGraphFile = config.getString(JOB_GRAPH_FILE_PATH, null);
+		String jobGraphFile = config.getString(JOB_GRAPH_FILE_PATH, "job.graph");
 		if (jobGraphFile != null) {
 			File fp = new File(jobGraphFile);
 			if (fp.isFile()) {
@@ -465,63 +268,11 @@ public class YarnFlinkApplicationMasterRunner implements LeaderContender, OnComp
 				input.close();
 			}
 		}
+		if (jg == null) {
+			throw new Exception("Fail to load job graph " + jobGraphFile);
+		}
 		return jg;
 	}
-
-	/**
-	 * 
-	 * @param baseDirectory
-	 * @param additional
-	 * 
-	 * @return The configuration to be used by the TaskExecutors.
-	 */
-	@SuppressWarnings("deprecation")
-	private static Configuration createConfiguration(String baseDirectory, Map<String, String> additional) {
-		LOG.info("Loading config from directory " + baseDirectory);
-
-		Configuration configuration = GlobalConfiguration.loadConfiguration(baseDirectory);
-
-		configuration.setString(ConfigConstants.FLINK_BASE_DIR_PATH_KEY, baseDirectory);
-
-		// add dynamic properties to JobManager configuration.
-		for (Map.Entry<String, String> property : additional.entrySet()) {
-			configuration.setString(property.getKey(), property.getValue());
-		}
-
-		// override zookeeper namespace with user cli argument (if provided)
-		String cliZKNamespace = ENV.get(YarnConfigKeys.ENV_ZOOKEEPER_NAMESPACE);
-		if (cliZKNamespace != null && !cliZKNamespace.isEmpty()) {
-			configuration.setString(HighAvailabilityOptions.HA_CLUSTER_ID, cliZKNamespace);
-		}
-
-		// if a web monitor shall be started, set the port to random binding
-		if (configuration.getInteger(ConfigConstants.JOB_MANAGER_WEB_PORT_KEY, 0) >= 0) {
-			configuration.setInteger(ConfigConstants.JOB_MANAGER_WEB_PORT_KEY, 0);
-		}
-
-		// if the user has set the deprecated YARN-specific config keys, we add the 
-		// corresponding generic config keys instead. that way, later code needs not
-		// deal with deprecated config keys
-
-		BootstrapTools.substituteDeprecatedConfigKey(configuration,
-			ConfigConstants.YARN_HEAP_CUTOFF_RATIO,
-			ConfigConstants.CONTAINERIZED_HEAP_CUTOFF_RATIO);
-
-		BootstrapTools.substituteDeprecatedConfigKey(configuration,
-			ConfigConstants.YARN_HEAP_CUTOFF_MIN,
-			ConfigConstants.CONTAINERIZED_HEAP_CUTOFF_MIN);
-
-		BootstrapTools.substituteDeprecatedConfigPrefix(configuration,
-			ConfigConstants.YARN_APPLICATION_MASTER_ENV_PREFIX,
-			ConfigConstants.CONTAINERIZED_MASTER_ENV_PREFIX);
-
-		BootstrapTools.substituteDeprecatedConfigPrefix(configuration,
-			ConfigConstants.YARN_TASK_MANAGER_ENV_PREFIX,
-			ConfigConstants.CONTAINERIZED_TASK_MANAGER_ENV_PREFIX);
-
-		return configuration;
-	}
-
 
 	//-------------------------------------------------------------------------------------
 	// Fatal error handler
@@ -534,54 +285,6 @@ public class YarnFlinkApplicationMasterRunner implements LeaderContender, OnComp
 		shutdown(ApplicationStatus.FAILED, exception.getMessage());
 	}
 
-	//----------------------------------------------------------------------------------------------
-	// Leadership methods
-	//----------------------------------------------------------------------------------------------
-
-	@Override
-	public void grantLeadership(final UUID leaderSessionID) {
-		synchronized (lock) {
-			if (!running) {
-				LOG.info("JobManagerRunner already shutdown.");
-				return;
-			}
-
-			LOG.info("JobMaster was granted leadership with session id {}.", leaderSessionID);
-
-			// The operation may be blocking, but since this runner is idle before it been granted the leadership,
-			// it's okay that job manager wait for the operation complete
-			jmLeaderElectionService.confirmLeaderSessionID(leaderSessionID);
-
-			
-			// Double check the leadership in case this one is failed and yarn start another one.
-			if (jmLeaderElectionService.hasLeadership()) {
-				try {
-					jobMaster.start(leaderSessionID);
-					LOG.debug("JobMaster started");
-				} catch (Exception e) {
-					onFatalError(new Exception("Could not start the job manager.", e));
-				}
-			} else {
-				LOG.info("There are another job master ");
-				jobFinishedByOther();
-			}
-		}
-	}
-
-	@Override
-	public void revokeLeadership() {
-		throw new UnsupportedOperationException("Yarn app master does not need leader select");
-	}
-
-	@Override
-	public String getAddress() {
-		return jobMaster.getAddress();
-	}
-
-	@Override
-	public void handleError(Exception exception) {
-		throw new UnsupportedOperationException("Yarn app master does not need leader select");
-	}
 	//----------------------------------------------------------------------------------------------
 	// Result and error handling methods
 	//----------------------------------------------------------------------------------------------
